+++ hob.cabal
@@ -41,6 +41,7 @@ library
                          Hob.Ui.Editor.Fancy.Types,
                          Hob.Ui.Editor.Fancy.Parser,
                          Hob.Ui.Editor.Fancy.Shapes,
+                         Hob.Ui.Editor.Fancy.Commands,
                          Hob.Command.FindText, Hob.Command.ReplaceText, Hob.Command.CloseCurrentTab,
                          Hob.Command.SaveCurrentTab, Hob.Command.NewTab, Hob.Command.FocusCommandEntry,
                          Hob.Command.FocusNextTab, Hob.Command.ReloadSidebarTree, Hob.Command.FocusSidebar,
